move OpTreeContext into support, like in scala-2
